requirements.txt
# Python requirements for DiscordSam bot and dependencies

# Note: torch install should be system-specific (see comment)

discord.py>=2.3.2
python-dotenv>=1.0.0
openai>=1.3.0
aiohttp>=3.8.0
pydub>=0.25.0

# HTML parsing fallback
beautifulsoup4>=4.12.2

# Whisper speech-to-text (choose one, usually openai-whisper)

openai-whisper>=20231117
torch>=2.1.0
torchaudio>=2.1.0


# Vector DB for RAG

chromadb>=0.4.22

# Automated browser for scraping, headless tasks

playwright>=1.30.0

youtube-transcript-api>=0.6.0
numpy
psutil>=5.9.0
dateparser>=1.1.8
